// old-v7/packages/core/src/lib/types.ts
export type NgtUnknownRecord = Record<string, unknown>;
// ...
export interface NgtIntersection extends THREE.Intersection {
  /** The event source (the object which registered the handler) */
  eventObject: THREE.Object3D;
}

export type NgtRenderer = {
  render: (scene: THREE.Scene, camera: THREE.Camera) => any;
};
export type NgtCamera = THREE.OrthographicCamera | THREE.PerspectiveCamera;

export type NgtDomEvent = PointerEvent | MouseEvent | WheelEvent;
export interface NgtIntersectionEvent<TSourceEvent> extends NgtIntersection {
  /** An array of intersections */
  intersections: NgtIntersection[];
  /** vec3.set(pointer.x, pointer.y, 0).unproject(camera) */
  unprojectedPoint: THREE.Vector3;
  /** Normalized event coordinates (x: [-1, 1], y: [-1, 1]) */
  pointer: THREE.Vector2;
  /** delta between first click and this event */
  delta: number;
  /** the ray casted from the camera */
  ray: THREE.Ray;
  /** the camera used to raycast */
  camera: NgtCamera;
  /** stopPropagation will stop underlying handlers from firing */
  stopPropagation: () => void;
  /** the original event */
  nativeEvent: TSourceEvent;
  /** if the event was stopped by calling stopPropagation */
  stopped: boolean;
}

export interface NgtPointerCaptureTarget {
  intersection: NgtIntersection;
  target: Element;
}

export type NgtThreeEvent<TEvent> = NgtIntersectionEvent<TEvent>;

export interface NgtEventHandlers {
  click?: (event: NgtThreeEvent<MouseEvent>) => void;
  contextmenu?: (event: NgtThreeEvent<MouseEvent>) => void;
  dblclick?: (event: NgtThreeEvent<MouseEvent>) => void;
  pointerup?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointerdown?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointerover?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointerout?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointerenter?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointerleave?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointermove?: (event: NgtThreeEvent<PointerEvent>) => void;
  pointermissed?: (event: MouseEvent) => void;
  pointercancel?: (event: NgtThreeEvent<PointerEvent>) => void;
  wheel?: (event: NgtThreeEvent<WheelEvent>) => void;
}
// ...
export interface NgtState {
  ready: boolean;
  /** Render loop flags */
  frameloop: 'always' | 'demand' | 'never';
  /* WebGLRenderer instance */
  gl: THREE.WebGLRenderer;
  /** Default camera */
  camera: NgtCamera & { manual?: boolean };
  cameraRef: NgtRef<NgtCamera & { manual?: boolean }>;
  /** Default scene */
  scene: THREE.Scene;
  sceneRef: NgtRef<THREE.Scene>;
  /** Normalized event coordinates */
  pointer: THREE.Vector2;
  /** Default raycaster */
  raycaster: THREE.Raycaster;
  /** Default clock */
  clock: THREE.Clock;
  /** Reactive pixel-size of the canvas */
  size: NgtSize;
  /** Reactive size of the viewport in threejs units */
  viewport: NgtViewport & {
    getCurrentViewport: (
      camera?: NgtCamera,
      target?: THREE.Vector3 | Parameters<THREE.Vector3['set']>,
      size?: NgtSize
    ) => Omit<NgtViewport, 'dpr' | 'initialDpr'>;
  };
  /** Adaptive performance interface */
  performance: NgtPerformance;
  /* Whether to enable r139's THREE.ColorManagement.legacyMode */
  legacy: boolean;
  /** Shortcut to gl.outputEncoding = LinearEncoding */
  linear: boolean;
  /** Shortcut to gl.toneMapping = NoTonemapping */
  flat: boolean;
  /** Event layer interface, contains the event handler and the node they're connected to */
  events: NgtEventManager<any>;
  /** Currently used controls */
  controls: THREE.EventDispatcher | null;
  /** XR interface */
  xr: { connect: () => void; disconnect: () => void };
  /** When the canvas was clicked but nothing was hit */
  onPointerMissed?: (event: MouseEvent) => void;
  /* internals */
  internal: NgtInternalState;
  /** If this state model is layerd (via createPortal) then this contains the previous layer */
  previousRoot?: NgtStateGetter;
  /** Flags the canvas for render, but doesn't render in itself */
  invalidate: (frames?: number) => void;
  /** Advance (render) one step */
  advance: (timestamp: number, runGlobalCallbacks?: boolean) => void;
  setPerformanceCurrent: (current: number) => void;
  setEvents: (events: Partial<NgtEventManager<any>>) => void;
  setFrameloop: (frameloop: NgtState['frameloop']) => void;
  setSize: (width: number, height: number, top?: number, left?: number, updateStyle?: boolean) => void;
  setDpr: (dpr: NgtDpr) => void;
  setCamera: (camera: NgtCamera) => void;
  addInteraction(interaction: THREE.Object3D): void;
  removeInteraction(uuid: string): void;
}

export type NgtStateGetter = () => NgtState;

export interface NgtRenderState extends NgtState {
  delta: number;
  frame?: XRFrame;
}

export interface NgtBeforeRender<TObject extends THREE.Object3D = THREE.Object3D> {
  state: NgtRenderState;
  object: TObject;
}

export type NgtBeforeRenderCallback<TObject> = (state: NgtRenderState, obj: TObject) => void;

export interface NgtBeforeRenderRecord {
  obj?: THREE.Object3D | NgtRef;
  callback: NgtBeforeRenderCallback<THREE.Object3D | undefined>;
  priority?: number;
}

export type NgtAttachFunction = (parent: NgtRef, child: NgtRef) => void | (() => void);

export interface NgtInstanceLocalState {
  /** the state getter of the canvas that the instance is being rendered to */
  stateGetter: NgtStateGetter;
  rootGetter: NgtStateGetter;
  // non-object 3d
  instancesRefs: NgtRef<NgtRef[]>;
  // object-3d
  objectsRefs: NgtRef<NgtRef[]>;
  parentRef: NgtRef<NgtInstanceNode> | null;
  isPrimitive?: boolean;
  eventCount: number;
  handlers: Partial<NgtEventHandlers>;
  args?: unknown[];
  attach?: string[] | (() => void);
  attachValue?: unknown;
  memoized?: NgtUnknownRecord;
}

export type NgtInstanceNode<TNode = any> = TNode & {
  __ngt__: NgtInstanceLocalState;
  [key: string]: any;
};

export type NgtPrepareInstanceFn<TInstance> = (
  instance: TInstance,
  prepareOptions?: {
    uuid?: string;
    parentStateGetter?: NgtStateGetter;
    rootStateGetter?: NgtStateGetter;
  }
) => NgtInstanceNode<TInstance>;